+++ src/components/levelset/LevelsToolbar/LevelsToolbar.tsx
@@ -1,21 +1,25 @@
 import { useStore } from "effector-react";
-import { FC, useMemo } from "react";
+import { FC, useCallback, useMemo } from "react";
 import { getDriverFormat } from "drivers";
 import {
   $currentBuffer,
+  $currentBufferHasOtherOpened,
   $currentDriverFormat,
   $currentDriverName,
   $currentLevel,
   appendLevel,
   closeCurrentLevel,
+  closeOtherLevels,
   deleteCurrentLevel,
   insertAtCurrentLevel,
 } from "models/levelsets";
 import { Button, ButtonDropdown, Toolbar } from "ui/button";
 import { ask } from "ui/feedback";
-import { svgs } from "ui/icon";
+import { IconStack, IconStackType, svgs } from "ui/icon";
 import { ColorType } from "ui/types";
 import { fmtLevelFull, fmtLevelNumber } from "../fmt";
+
+const closeOtherStack: IconStack = [[IconStackType.Index, <svgs.Cross />]];
 
 interface Props {
   isCompact?: boolean;
@@ -32,6 +36,8 @@ export const LevelsToolbar: FC<Props> = ({ isCompact = false }) => {
 
   const levelsCount = levelset.levels.length;
   const levelsCountDigits = String(levelsCount).length;
+
+  const hasOtherOpened = useStore($currentBufferHasOtherOpened);
 
   const levelFullReference =
     level &&
@@ -109,6 +115,41 @@ export const LevelsToolbar: FC<Props> = ({ isCompact = false }) => {
     levelsCountDigits,
   )}`;
 
+  const closeButton = (
+    <Button
+      icon={<svgs.Cross />}
+      disabled={!level}
+      onClick={closeCurrentLevel}
+      title={
+        levelFullReference ? `Close level tab (${levelFullReference})` : ""
+      }
+    />
+  );
+  const closeOthersTitle = "Close other levels tabs";
+  const handleCloseOthers = useCallback(async () => {
+    if (
+      await ask(
+        <>
+          Are you sure you want to close ALL OTHER levels BUT "
+          <b>{levelFullReference}</b>"?
+        </>,
+        {
+          buttons: {
+            okText: <>Close OTHER BUT "{levelFullReference}"</>,
+            ok: {
+              autoFocus: false,
+            },
+            cancel: {
+              autoFocus: true,
+            },
+          },
+        },
+      )
+    ) {
+      closeOtherLevels();
+    }
+  }, [levelFullReference]);
+
   return (
     <>
       {cannotAddLevel ||
@@ -142,14 +183,32 @@ export const LevelsToolbar: FC<Props> = ({ isCompact = false }) => {
           }
         />
       )}
-      <Button
-        icon={<svgs.Cross />}
-        disabled={!level}
-        onClick={closeCurrentLevel}
-        title={
-          levelFullReference ? `Close level tab (${levelFullReference})` : ""
-        }
-      />
+
+      {hasOtherOpened && !isCompact ? (
+        <ButtonDropdown standalone={closeButton}>
+          <Toolbar>
+            <Button
+              icon={<svgs.Cross />}
+              iconStack={closeOtherStack}
+              onClick={handleCloseOthers}
+            >
+              {closeOthersTitle}
+            </Button>
+          </Toolbar>
+        </ButtonDropdown>
+      ) : (
+        <>
+          {closeButton}
+          {hasOtherOpened && (
+            <Button
+              icon={<svgs.Cross />}
+              iconStack={closeOtherStack}
+              onClick={handleCloseOthers}
+              title={closeOthersTitle}
+            />
+          )}
+        </>
+      )}
     </>
   );
 };